requirements.txt
openai>=1.40.0
httpx>=0.27.0
pandas>=2.2.0
pyyaml>=6.0.1
tqdm>=4.66.0
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.5.0
